pyproject.toml
[tool.poetry]
name = "strange-attractors"
version = "0.1.0"
description = ""
authors = ["itz-amethyst <pransermi@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.108.0"
aiocache = "0.12.2"
annotated-types = "0.6.0"
anyio = "4.2.0"
certifi = "2023.11.17"
charset-normalizer = "3.3.2"
click = "8.1.7"
cloudpickle = "3.0.0"
colorcet = "3.0.1"
dask = "2023.12.1"
datashader = "0.16.0"
fsspec = "2023.12.2"
h11 = "0.14.0"
idna = "3.6"
importlib-metadata = "7.0.1"
llvmlite = "0.41.1"
locket = "1.0.0"
loguru = "0.7.2"
multipledispatch = "1.0.0"
numba = "0.58.1"
numpy = "1.26.3"
orjson = "3.9.10"
packaging = "23.2"
pandas = "2.1.4"
param = "2.0.1"
partd = "1.4.1"
pillow = "10.2.0"
prometheus-client = "0.19.0"
prometheus-fastapi-instrumentator = "6.1.0"
pyct = "0.5.0"
pydantic = "2.5.3"
pydantic-settings = "2.1.0"
pydantic-core = "2.14.6"
python-dateutil = "2.8.2"
python-dotenv = "1.0.0"
pytz = "2023.3.post1"
pyyaml = "6.0.1"
requests = "2.31.0"
scipy = "1.11.4"
six = "1.16.0"
sniffio = "1.3.0"
starlette = "0.32.0.post1"
toolz = "0.12.0"
typing-extensions = "4.9.0"
tzdata = "2023.4"
urllib3 = "2.1.0"
uvicorn = "0.25.0"
xarray = "2023.12.0"
zipp = "3.17.0"
structlog = "^24.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
